dbtr/server/requirements.txt
fastapi
uvicorn
snowflake-id
skaff-telemetry>=0.1,<1
humanize
dbt-core
pandas
